+++ src/pages/components/landing-page/about.tsx
