+++ src/pages/components/landing-page/footer.tsx
