+++ src/pages/components/landing-page/hero.tsx
