+++ src/pages/components/landing-page/profiles/profile2.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Avatar from '@material-ui/core/Avatar';
 import Typography from '@material-ui/core/Typography';
-import sengoku from '../../../assets/img/sengoku-one-piece.jpg'
+import sengoku from '../../../../assets/img/sengoku-one-piece.jpg'
 import Image from 'next/image';
 
 const useStyles = makeStyles((theme) => ({
