+++ src/pages/components/landing-page/register.tsx
