--- src/pages/main/dashboard.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const dashboard = () => {
-  return (
-    <div>dashboard</div>
-  )
-}
-
-export default dashboard
